requirements-test.txt
pytest==8.3.5
pytest-django==4.7.0
pytest-qt==4.4.0
pytest-mock==3.14.0
pytest-cov==4.1.0
faker==22.0.0